+++ _macros.php
@@ -1,28 +1,26 @@
 <?php namespace ProcessWire;
 
-wire($this->api_var)->_addMacro = array(
 
-    'iff' => array(
-        'iff',
-		'if (!empty(%node.word) && %node.word) { $x = %node.word;',
-        '}'
-    ),
+$view->addMacro(
+    'iff',
+    'if (!empty(%node.word) && %node.word) { $x = %node.word;',
+    '}'
+);
 
-    'page' => array(
-        'page',
-        '$p = \ProcessWire\wire("pages")->get(%node.word)',
-        ';'
-    ),
+$view->addMacro(
+    'page',
+    '$p = \ProcessWire\wire("pages")->get(%node.word)',
+    ';'
+);
 
-    'pages' => array(
-        'pages',
-        '$pArr = \ProcessWire\wire("pages")->find(%node.word)',
-        ';'
-    ),
+$view->addMacro(
+    'pages',
+    '$pArr = \ProcessWire\wire("pages")->find(%node.word)',
+    ';'
+);
 
-    'setvar' => array(
-        'setvar',
-        '$vars = %node.array; ${$vars[0]}=$vars[1]',
-        ';'
-    )
+$view->addMacro(
+    'setvar',
+    '$vars = %node.array; ${$vars[0]}=$vars[1]',
+    ';'
 );